backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Redis
redis==5.0.1
hiredis==2.2.3

# Storage (MinIO/S3)
minio==7.2.0
boto3==1.29.7

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ML/AI
numpy<2.0.0
torch==2.1.0
transformers==4.35.2
sentence-transformers==2.7.0
huggingface-hub>=0.20.0
accelerate==0.25.0

# Document processing
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.10.3
pypandoc==1.12

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Logging
loguru==0.7.2
